pyproject.toml
[project]
name = "ameliavoice"
version = "0.1.0"
description = "Voice conversation app - Phase 1: Google ASR and TTS"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "python-multipart>=0.0.6",
    "openai>=1.0.0",
    "google-cloud-speech>=2.24.0",
    "google-cloud-texttospeech>=2.16.0",
    "pydub>=0.25.1",
    "twilio>=9.0.0",
    "tavily-python>=0.5.0",
]

[project.optional-dependencies]
dev = ["pytest", "httpx"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["backend"]

[tool.uv]
dev-dependencies = []
